Updated CSS File
I changed the header, "Vision", and "Who We Need" fonts. I was going for a futura bold but couldn't find that. Think Supreme. I'll keep searching. I also reduced the space between the header and the top of the browser. And of course I changed the colors of the background for the header and the main text to white instead of grey. I also found a way to reduce the space between the header text and the X. Let me know what you guys think. Updated CSS file on Github.